requirements.txt
streamlit
mediapipe
opencv-python
Pillow
numpy
